Move new ConcludableTest tests to UnifierTest
