+++ logic/resolvable/ConcludableTest.java
@@ -312,378 +312,4 @@ public class ConcludableTest {
             assertTrue(hasConcludable.has().attribute().isa().isPresent());
         });
     }
-
-    @Test
-    public void test_constraints_involving_a_variable_are_always_satisfied() {
-        Pair<String, String>[] constraintValuePairs = new Pair[]{
-                // Both var
-                new Pair("{ $x has attr $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr = $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr != $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr <= $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr >= $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr < $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr > $v; }", "{$y has attr $w;}"),
-
-                // Boolean
-                new Pair("{ $x has attr $v; }", "{$y has attr true;}"),
-                new Pair("{ $x has attr $v; }", "{$y has attr false;}"),
-
-                new Pair("{ $x has attr = $v; }", "{$y has attr true;}"),
-                new Pair("{ $x has attr = $v; }", "{$y has attr false;}"),
-
-                new Pair("{ $x has attr != $v; }", "{$y has attr true;}"),
-                new Pair("{ $x has attr != $v; }", "{$y has attr false;}"),
-
-                // Numeric assign, equality, gte, lte
-                new Pair("{ $x has attr $v; }", "{$y has attr 1;}"),
-                new Pair("{ $x has attr $v; }", "{$y has attr 1.0;}"),
-
-                new Pair("{ $x has attr = $v; }", "{$y has attr 1;}"),
-                new Pair("{ $x has attr = $v; }", "{$y has attr 1.0;}"),
-
-                new Pair("{ $x has attr >= $v; }", "{$y has attr 1;}"),
-                new Pair("{ $x has attr >= $v; }", "{$y has attr 1.0;}"),
-
-                new Pair("{ $x has attr <= $v; }", "{$y has attr 1;}"),
-                new Pair("{ $x has attr <= $v; }", "{$y has attr 1.0;}"),
-
-                // Numeric inequality, gt, lt
-                new Pair("{ $x has attr != $v; }", "{$y has attr 1;}"),
-                new Pair("{ $x has attr != $v; }", "{$y has attr 1.0;}"),
-
-                new Pair("{ $x has attr > $v; }", "{$y has attr 1;}"),
-                new Pair("{ $x has attr > $v; }", "{$y has attr 1.0;}"),
-
-                new Pair("{ $x has attr < $v; }", "{$y has attr 1;}"),
-                new Pair("{ $x has attr < $v; }", "{$y has attr 1.0;}"),
-
-                // String comparisons
-                new Pair("{ $x has attr = $v; }", "{$y has attr \"one\";}"),
-                new Pair("{ $x has attr >= $v; }", "{$y has attr \"one\";}"),
-                new Pair("{ $x has attr <= $v; }", "{$y has attr \"one\";}"),
-                new Pair("{ $x has attr != $v; }", "{$y has attr \"one\";}"),
-                new Pair("{ $x has attr > $v; }", "{$y has attr \"one\";}"),
-                new Pair("{ $x has attr < $v; }", "{$y has attr \"one\";}"),
-
-                // DateTime
-                new Pair("{ $x has attr $v; }", "{$y has attr 2022-01-01;}"),
-                new Pair("{ $x has attr = $v; }", "{$y has attr 2022-01-01;}"),
-                new Pair("{ $x has attr != $v; }", "{$y has attr 2022-01-01;}"),
-                new Pair("{ $x has attr >= $v; }", "{$y has attr 2022-01-01;}"),
-                new Pair("{ $x has attr <= $v; }", "{$y has attr 2022-01-01;}"),
-                new Pair("{ $x has attr > $v; }", "{$y has attr 2022-01-01;}"),
-                new Pair("{ $x has attr < $v; }", "{$y has attr 2022-01-01;}"),
-        };
-
-        for (Pair<String, String> pair : constraintValuePairs) {
-            Concludable constraint = iterate(Concludable.create(parseConjunction(pair.first()))).next();
-            Concludable value = iterate(Concludable.create(parseConjunction(pair.second()))).next();
-
-            assertTrue(constraint.isHas());
-            assertTrue(value.isHas());
-            assertTrue(Unifier.Builder.unificationSatisfiable(constraint.asHas().attribute(), value.asHas().attribute()));
-        }
-    }
-    @Test
-    public void test_variables_always_satisfy_value_constraint() {
-        Pair<String, String>[] constraintValuePairs = new Pair[]{
-                // Both var
-                new Pair("{ $x has attr $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr = $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr != $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr <= $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr >= $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr < $v; }", "{$y has attr $w;}"),
-                new Pair("{ $x has attr > $v; }", "{$y has attr $w;}"),
-
-                // Boolean
-                new Pair("{ $x has attr true; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr false; }", "{$y has attr $v;}"),
-
-                new Pair("{ $x has attr = true; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr = false; }", "{$y has attr $v;}"),
-
-                new Pair("{ $x has attr != true; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr != false; }", "{$y has attr $v;}"),
-
-                // Numeric assign, equality, gte, lte
-                new Pair("{ $x has attr 1; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr 1.0; }", "{$y has attr $v;}"),
-
-                new Pair("{ $x has attr = 1; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr = 1.0; }", "{$y has attr $v;}"),
-
-                new Pair("{ $x has attr >= 1; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr >= 1.0; }", "{$y has attr $v;}"),
-
-                new Pair("{ $x has attr <= 1; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr <= 1.0; }", "{$y has attr $v;}"),
-
-                // Numeric inequality, gt, lt
-                new Pair("{ $x has attr != 1; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr != 1.0; }", "{$y has attr $v;}"),
-
-                new Pair("{ $x has attr > 1; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr > 1.0; }", "{$y has attr $v;}"),
-
-                new Pair("{ $x has attr < 1; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr < 1.0; }", "{$y has attr $v;}"),
-
-                // String comparisons
-                new Pair("{ $x has attr \"one\"; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr >= \"one\"; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr <= \"one\"; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr != \"one\"; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr > \"one\"; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr < \"one\"; }", "{$y has attr $v;}"),
-
-                // DateTime
-                new Pair("{ $x has attr 2022-01-01; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr = 2022-01-01; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr != 2022-01-01; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr >= 2022-01-01; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr <= 2022-01-01; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr > 2022-01-01; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr < 2022-01-01; }", "{$y has attr $v;}"),
-
-                // Substring
-                new Pair("{ $x has attr like \"\"; }", "{$y has attr $v;}"),
-                new Pair("{ $x has attr contains \"jan\"; }", "{$y has attr $v;}"),
-        };
-
-        for (Pair<String, String> pair : constraintValuePairs) {
-            Concludable constraint = iterate(Concludable.create(parseConjunction(pair.first()))).next();
-            Concludable value = iterate(Concludable.create(parseConjunction(pair.second()))).next();
-
-            assertTrue(constraint.isHas());
-            assertTrue(value.isHas());
-            assertTrue(Unifier.Builder.unificationSatisfiable(constraint.asHas().attribute(), value.asHas().attribute()));
-        }
-    }
-
-    @Test
-    public void test_constants_which_satisfy_value_constraint() {
-        Pair<String, String>[] constraintValuePairs = new Pair[]{
-                // Boolean
-                new Pair("{ $x true; }", "{$y true;}"),
-                new Pair("{ $x false; }", "{$y false;}"),
-                new Pair("{ $x = true; }", "{$y true;}"),
-                new Pair("{ $x = false; }", "{$y false;}"),
-                new Pair("{ $x != true; }", "{$y false;}"),
-                new Pair("{ $x != false; }", "{$y true;}"),
-
-                // Numeric assign, equality, gte, lte
-                new Pair("{ $x 1; }", "{$y 1;}"),
-                new Pair("{ $x 1; }", "{$y 1.0;}"),
-                new Pair("{ $x 1.0; }", "{$y 1;}"),
-                new Pair("{ $x 1.0; }", "{$y 1.0;}"),
-
-                new Pair("{ $x = 1; }", "{$y 1;}"),
-                new Pair("{ $x = 1; }", "{$y 1.0;}"),
-                new Pair("{ $x = 1.0; }", "{$y 1;}"),
-                new Pair("{ $x = 1.0; }", "{$y 1.0;}"),
-
-                new Pair("{ $x >= 1; }", "{$y 1;}"),
-                new Pair("{ $x >= 1; }", "{$y 1.0;}"),
-                new Pair("{ $x >= 1.0; }", "{$y 1;}"),
-                new Pair("{ $x >= 1.0; }", "{$y 1.0;}"),
-
-                new Pair("{ $x <= 1; }", "{$y 1;}"),
-                new Pair("{ $x <= 1; }", "{$y 1.0;}"),
-                new Pair("{ $x <= 1.0; }", "{$y 1;}"),
-                new Pair("{ $x <= 1.0; }", "{$y 1.0;}"),
-
-                // Numeric inequality, gt, lt
-                new Pair("{ $x != 1; }", "{$y 2;}"),
-                new Pair("{ $x != 1; }", "{$y 2.0;}"),
-                new Pair("{ $x != 1.0; }", "{$y 2;}"),
-                new Pair("{ $x != 1.0; }", "{$y 2.0;}"),
-
-                new Pair("{ $x > 1; }", "{$y 2;}"),
-                new Pair("{ $x > 1; }", "{$y 2.0;}"),
-                new Pair("{ $x > 1.0; }", "{$y 2;}"),
-                new Pair("{ $x > 1.0; }", "{$y 2.0;}"),
-
-                new Pair("{ $x < 1; }", "{$y -2;}"),
-                new Pair("{ $x < 1; }", "{$y -2.0;}"),
-                new Pair("{ $x < 1.0; }", "{$y -2;}"),
-                new Pair("{ $x < 1.0; }", "{$y -2.0;}"),
-
-                // String comparisons
-                new Pair("{ $x \"one\"; }", "{$y \"one\";}"),
-                new Pair("{ $x = \"one\"; }", "{$y \"one\";}"),
-                new Pair("{ $x >= \"one\"; }", "{$y \"one\";}"),
-                new Pair("{ $x <= \"one\"; }", "{$y \"one\";}"),
-                new Pair("{ $x != \"one\"; }", "{$y \"two\";}"),
-                new Pair("{ $x > \"one\"; }", "{$y \"two\";}"),
-                new Pair("{ $x < \"two\"; }", "{$y \"one\";}"),
-
-                // String comparisons, case-sensitivity
-                new Pair("{ $x != \"one\"; }", "{$y \"ONE\";}"),
-                new Pair("{ $x > \"TWO\"; }", "{$y \"one\";}"),
-                new Pair("{ $x < \"one\"; }", "{$y \"TWO\";}"),
-
-                // DateTime
-                new Pair("{ $x 2022-01-01; }", "{$y 2022-01-01;}"),
-                new Pair("{ $x = 2022-01-01; }", "{$y  2022-01-01;}"),
-                new Pair("{ $x != 2022-01-01; }", "{$y  2022-02-02;}"),
-
-                new Pair("{ $x >= 2022-01-01; }", "{$y 2022-01-01;}"),
-                new Pair("{ $x >= 2022-01-01; }", "{$y 2022-02-02;}"),
-                new Pair("{ $x <= 2022-01-01; }", "{$y 2022-01-01;}"),
-                new Pair("{ $x <= 2022-02-02; }", "{$y 2022-01-01;}"),
-
-                new Pair("{ $x > 2022-01-01; }", "{$y 2022-02-02;}"),
-                new Pair("{ $x < 2022-02-02; }", "{$y 2022-01-01;}"),
-        };
-
-        for (Pair<String, String> pair : constraintValuePairs) {
-            Concludable constraint = iterate(Concludable.create(parseConjunction(pair.first()))).next();
-            Concludable value = iterate(Concludable.create(parseConjunction(pair.second()))).next();
-
-            assertTrue(constraint.isAttribute());
-            assertTrue(value.isAttribute());
-            assertTrue(Unifier.Builder.unificationSatisfiable(constraint.asAttribute().attribute(), value.asAttribute().attribute()));
-        }
-    }
-
-    @Test
-    public void test_constants_which_dont_satisfy_value_constraint() {
-        Pair<String, String>[] constraintValuePairs = new Pair[]{
-                // Boolean
-                new Pair("{ $x true; }", "{$y false;}"),
-                new Pair("{ $x false; }", "{$y true;}"),
-                new Pair("{ $x = true; }", "{$y false;}"),
-                new Pair("{ $x = false; }", "{$y true;}"),
-                new Pair("{ $x != true; }", "{$y true;}"),
-                new Pair("{ $x != false; }", "{$y false;}"),
-
-                // Numeric assign, equality, gte, lte
-                new Pair("{ $x 1; }", "{$y 2;}"),
-                new Pair("{ $x 1; }", "{$y 2.0;}"),
-                new Pair("{ $x 1.0; }", "{$y 2;}"),
-                new Pair("{ $x 1.0; }", "{$y 2.0;}"),
-
-                new Pair("{ $x = 1; }", "{$y 2;}"),
-                new Pair("{ $x = 1; }", "{$y 2.0;}"),
-                new Pair("{ $x = 1.0; }", "{$y 2;}"),
-                new Pair("{ $x = 1.0; }", "{$y 2.0;}"),
-
-                new Pair("{ $x = 1; }", "{$y -1;}"),
-                new Pair("{ $x = 1; }", "{$y -1.0;}"),
-                new Pair("{ $x = 1.0; }", "{$y -1;}"),
-                new Pair("{ $x = 1.0; }", "{$y -1.0;}"),
-
-                new Pair("{ $x >= 1; }", "{$y -1;}"),
-                new Pair("{ $x >= 1; }", "{$y -1.0;}"),
-                new Pair("{ $x >= 1.0; }", "{$y -1;}"),
-                new Pair("{ $x >= 1.0; }", "{$y -1.0;}"),
-
-                new Pair("{ $x <= 1; }", "{$y 2;}"),
-                new Pair("{ $x <= 1; }", "{$y 2.0;}"),
-                new Pair("{ $x <= 1.0; }", "{$y 2;}"),
-                new Pair("{ $x <= 1.0; }", "{$y 2.0;}"),
-
-                // Numeric inequality, gt, lt
-                new Pair("{ $x != 1; }", "{$y 1;}"),
-                new Pair("{ $x != 1; }", "{$y 1.0;}"),
-                new Pair("{ $x != 1.0; }", "{$y 1;}"),
-                new Pair("{ $x != 1.0; }", "{$y 1.0;}"),
-
-                new Pair("{ $x > 1; }", "{$y -1;}"),
-                new Pair("{ $x > 1; }", "{$y -1.0;}"),
-                new Pair("{ $x > 1.0; }", "{$y -1;}"),
-                new Pair("{ $x > 1.0; }", "{$y -1.0;}"),
-
-                new Pair("{ $x > 1; }", "{$y 1;}"),
-                new Pair("{ $x > 1; }", "{$y 1.0;}"),
-                new Pair("{ $x > 1.0; }", "{$y 1;}"),
-                new Pair("{ $x > 1.0; }", "{$y 1.0;}"),
-
-                new Pair("{ $x < 1; }", "{$y 1;}"),
-                new Pair("{ $x < 1; }", "{$y 1.0;}"),
-                new Pair("{ $x < 1.0; }", "{$y 1;}"),
-                new Pair("{ $x < 1.0; }", "{$y 1.0;}"),
-
-                new Pair("{ $x < 1; }", "{$y 2;}"),
-                new Pair("{ $x < 1; }", "{$y 2.0;}"),
-                new Pair("{ $x < 1.0; }", "{$y 2;}"),
-                new Pair("{ $x < 1.0; }", "{$y 2.0;}"),
-
-                // String comparisons
-                new Pair("{ $x \"one\"; }", "{$y \"two\";}"),
-                new Pair("{ $x = \"one\"; }", "{$y \"two\";}"),
-                new Pair("{ $x >= \"two\"; }", "{$y \"one\";}"),
-                new Pair("{ $x <= \"one\"; }", "{$y \"two\";}"),
-                new Pair("{ $x != \"one\"; }", "{$y \"one\";}"),
-                new Pair("{ $x > \"one\"; }", "{$y \"one\";}"),
-                new Pair("{ $x > \"two\"; }", "{$y \"one\";}"),
-                new Pair("{ $x < \"two\"; }", "{$y \"two\";}"),
-                new Pair("{ $x < \"one\"; }", "{$y \"two\";}"),
-
-                // String comparisons, case-sensitivity
-                new Pair("{ $x \"OnE\"; }", "{$y \"oNe\";}"),
-                new Pair("{ $x = \"OnE\"; }", "{$y \"oNe\";}"),
-                new Pair("{ $x > \"two\"; }", "{$y \"ONE\";}"),
-                new Pair("{ $x < \"ONE\"; }", "{$y \"two\";}"),
-
-                // DateTime
-                new Pair("{ $x 2022-01-01; }", "{$y  2022-02-02;}"),
-                new Pair("{ $x = 2022-01-01; }", "{$y  2022-02-02;}"),
-                new Pair("{ $x != 2022-01-01; }", "{$y  2022-01-01;}"),
-
-                new Pair("{ $x >= 2022-02-02; }", "{$y 2022-01-01;}"),
-                new Pair("{ $x <= 2022-01-01; }", "{$y 2022-02-02;}"),
-
-                new Pair("{ $x > 2022-01-01; }", "{$y 2022-01-01;}"),
-                new Pair("{ $x > 2022-02-02; }", "{$y 2022-01-01;}"),
-                new Pair("{ $x < 2022-01-01; }", "{$y 2022-01-01;}"),
-                new Pair("{ $x < 2022-01-01; }", "{$y 2022-02-02;}"),
-        };
-
-        for (Pair<String, String> pair : constraintValuePairs) {
-            Concludable constraint = iterate(Concludable.create(parseConjunction(pair.first()))).next();
-            Concludable value = iterate(Concludable.create(parseConjunction(pair.second()))).next();
-
-            assertTrue(constraint.isAttribute());
-            assertTrue(value.isAttribute());
-            assertFalse(Unifier.Builder.unificationSatisfiable(constraint.asAttribute().attribute(), value.asAttribute().attribute()));
-        }
-    }
-
-    @Test
-    public void test_constants_which_satisfy_substring_constraint() {
-        Pair<String, String>[] valueConstraintPairs = new Pair[]{
-                new Pair("{ $x has attr like \"[0-9]{2}-[a-z]{3}-[0-9]{4}\"; }", "{$y has attr \"01-jan-2022\";}"),
-                new Pair("{ $x has attr contains \"jan\"; }", "{$y has attr \"01-jan-2022\";}"),
-        };
-
-        for (Pair<String, String> pair : valueConstraintPairs) {
-            Concludable constraint = iterate(Concludable.create(parseConjunction(pair.first()))).next();
-            Concludable value = iterate(Concludable.create(parseConjunction(pair.second()))).next();
-
-            assertTrue(constraint.isHas());
-            assertTrue(value.isHas());
-            assertTrue(Unifier.Builder.unificationSatisfiable(constraint.asHas().attribute(), value.asHas().attribute()));
-        }
-    }
-
-    @Test
-    public void test_constants_which_dont_satisfy_substring_constraint() {
-        Pair<String, String>[] valueConstraintPairs = new Pair[]{
-                // Both constant
-                new Pair("{ $x has attr like \"[0-9]{2}-[a-z]{3}-[0-9]{4}\"; }", "{$y has attr \"01-01-2022\";}"),
-                new Pair("{ $x has attr contains \"jan\"; }", "{$y has attr \"01-feb-2022\";}"),
-        };
-
-        for (Pair<String, String> pair : valueConstraintPairs) {
-            Concludable constraint = iterate(Concludable.create(parseConjunction(pair.first()))).next();
-            Concludable value = iterate(Concludable.create(parseConjunction(pair.second()))).next();
-
-            assertTrue(constraint.isHas());
-            assertTrue(value.isHas());
-            assertFalse(Unifier.Builder.unificationSatisfiable(constraint.asHas().attribute(), value.asHas().attribute()));
-        }
-    }
 }
